#  perf: Improved diff reduce function from quadratic to linear.