## Benchmark

### Before
```
goos: darwin
goarch: amd64
pkg: github.com/mbtproject/mbt/lib
BenchmarkReduceToDiff10-8      	     300	        3934437 ns/op
BenchmarkReduceToDiff100-8     	      50	      36742533 ns/op
BenchmarkReduceToDiff1000-8    	       2	     528151540 ns/op
BenchmarkReduceToDiff10000-8   	       1	24740225433 ns/op
```

### After
```
goos: darwin
goarch: amd64
pkg: github.com/mbtproject/mbt/lib
BenchmarkReduceToDiff10-8      	     300	      4387785 ns/op
BenchmarkReduceToDiff100-8     	      50	     34644111 ns/op
BenchmarkReduceToDiff1000-8    	       3	 355000553 ns/op
BenchmarkReduceToDiff10000-8   	       1	3752071559 ns/op
PASS
```